Tweaked leaderboard output format

+++ .vshistory/LeaderboardEntry.cs/2025-05-09_20_12_24_270.cs
@@ -81,7 +81,7 @@ namespace Rashed_Blackjack
                 }
                 else
                 {
-                    return (double)Wins / (double)matchesPlayed;
+                    return Wins / matchesPlayed;
                 }
             }
         }
